backend/requirements.txt
fastapi
uvicorn
pydantic
tldextract
requests
google-generativeai
firebase-admin
google-cloud-firestore
python-dotenv
